helikon-opera_bot/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==26.1.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
choicelib==0.1.5
colorama==0.4.6
exceptiongroup==1.3.0
frozenlist==1.8.0
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
lxml==6.0.2
msgspec==0.19.0
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
python-telegram-bot==22.5
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
vbml==1.1.post1
vkbottle==4.7.0
vkbottle-types==5.199.99.12
yarl==1.23.0
